New Benchmark Maintenance

+++ app/controllers/app_maintenance/ifa_controller.rb
@@ -70,10 +70,12 @@ class AppMaintenance::IfaController < ApplicationController
   def strand_destroy
     strands
     current_strand
-    if @current_strand.destroy
-      flash[:notice] = "Strand Destroyed"
-      strands
-      @current_strand = @strands.first rescue nil
+    if false
+      if @current_strand.destroy
+        flash[:notice] = "Strand Destroyed"
+        strands
+        @current_strand = @strands.first rescue nil
+      end
     end
     render :partial =>  "edit_strand", :locals=>{:strand => @current_strand}
   end
@@ -118,6 +120,92 @@ class AppMaintenance::IfaController < ApplicationController
     render :partial =>  "edit_level", :locals=>{:level => @current_level}
   end
 
+  def benchmark_refresh
+    get_strand
+    levels
+    render :partial =>  "manage_benchmarks_strand", :locals=>{:strand => @current_strand}
+  end
+
+  def benchmark_toggle
+    current_benchmark
+    @current_benchmark.update_attributes(:is_active=> !@current_benchmark.is_active)
+    render :partial =>  "show_benchmark",  :locals=>{:strand => @current_benchmark.strand, :level => @current_benchmark.mastery_level,
+                                                     :benchmark => @current_benchmark, :benchmark_id => @current_benchmark.public_id}
+  end
+
+  def benchmark_destroy
+    current_benchmark
+    strand = @current_benchmark.strand
+    level = @current_benchmark.mastery_level
+    benchmark_id = @current_benchmark.public_id
+    if @current_benchmark.destroy
+      flash[:notice] = "Benchmark Destroyed"
+    end
+    render :partial =>  "show_benchmark",  :locals=>{:strand => strand, :level => level, :benchmark => nil, :benchmark_id => benchmark_id}
+  end
+
+  def benchmark_add
+    get_strand
+    get_level
+    @bench_types = @current_level.standard.act_bench_types
+    @current_benchmark = ActBench.new
+    set_default_bench_type(@current_benchmark)
+  end
+
+  def benchmark_edit
+    get_strand
+    get_level
+    current_benchmark
+    @bench_types = @current_level.standard.act_bench_types
+  end
+
+  def benchmark_create
+    get_strand
+    get_level
+    @bench_types = @current_level.standard.act_bench_types
+    @current_benchmark = ActBench.new
+    @current_benchmark.act_master_id = @current_level.standard.id
+    @current_benchmark.act_subject_id = @current_level.subject_area.id
+    @current_benchmark.act_standard_id = @current_strand.id
+    @current_benchmark.act_score_range_id = @current_level.id
+    @current_benchmark.user_id = @current_user.id
+    @current_benchmark.organization_id = @current_organization.id
+    @current_benchmark.description = params[:act_bench][:description]
+    if params[:act_bench][:act_benchmark_type_id] == ''
+      set_default_bench_type(@current_benchmark)
+    else
+      @current_benchmark.act_bench_type_id = params[:act_bench][:act_benchmark_type_id]
+    end
+    @current_benchmark.pos = params[:act_bench][:pos]
+    if @current_benchmark.save
+      flash[:notice] = "Benchmark Created"
+      @current_benchmark = ActBench.new
+    else
+      flash[:error] = @current_benchmark.errors.full_messages.to_sentence
+    end
+    render :benchmark_add
+  end
+
+  def benchmark_update
+    get_strand
+    get_level
+    current_benchmark
+    @bench_types = @current_level.standard.act_bench_types
+    @current_benchmark.user_id = @current_user.id
+    @current_benchmark.organization_id = @current_organization.id
+    @current_benchmark.description = params[:act_bench][:description]
+    if params[:act_bench][:act_benchmark_type_id] != ''
+      @current_benchmark.act_bench_type_id = params[:act_bench][:act_benchmark_type_id]
+    end
+    @current_benchmark.pos = params[:act_bench][:pos]
+    if @current_benchmark.save
+      flash[:notice] = "Benchmark Updated"
+    else
+      flash[:error] = @current_benchmark.errors.full_messages.to_sentence
+    end
+    render :benchmark_edit
+  end
+
   private
 
   def strand_params
@@ -150,11 +238,16 @@ class AppMaintenance::IfaController < ApplicationController
 
   def current_strand
     if !@strands.empty?
-      if params[:act_standard_id]
-        @current_strand = ActStandard.find_by_id(params[:act_standard_id]) rescue @strands.first
-      else
-        @current_strand = @strands.first
-      end
+      get_strand
+      @current_strand = @current_strand.nil? ? @strands.first : @current_strand
+    else
+      @current_strand = nil
+    end
+  end
+
+  def get_strand
+    if params[:act_standard_id]
+      @current_strand = ActStandard.find_by_id(params[:act_standard_id]) rescue nil
     else
       @current_strand = nil
     end
@@ -166,13 +259,28 @@ class AppMaintenance::IfaController < ApplicationController
 
   def current_level
     if !@levels.empty?
-      if params[:act_score_range_id]
-        @current_level = ActScoreRange.find_by_id(params[:act_score_range_id])
-      else
-        @current_level = @levels.first
-      end
+      get_level
+      @current_level = @current_level.nil? ? @levels.first : @current_level
     else
       @current_level = nil
     end
+  end
+
+  def get_level
+    if params[:act_score_range_id]
+      @current_level = ActScoreRange.find_by_id(params[:act_score_range_id]) rescue nil
+    else
+      @current_level = nil
+    end
+  end
+
+  def current_benchmark
+    if params[:act_benchmark_id]
+      @current_benchmark = ActBench.find_by_id(params[:act_benchmark_id]) rescue nil
+    end
+  end
+
+  def set_default_bench_type(benchmark)
+    benchmark.act_bench_type_id = @bench_types.empty? ? nil : @bench_types.first.id
   end
 end
